deprecated dui from rhino
